[Refactor] Extract module AutocorrectUnlessChangingAST
