+++ lib/rubocop.rb
@@ -47,6 +47,7 @@ require 'rubocop/cop/variable_inspector/variable_table'
 require 'rubocop/cop/mixin/annotation_comment'
 require 'rubocop/cop/mixin/array_syntax'
 require 'rubocop/cop/mixin/autocorrect_alignment'
+require 'rubocop/cop/mixin/autocorrect_unless_changing_ast'
 require 'rubocop/cop/mixin/check_assignment'
 require 'rubocop/cop/mixin/check_methods'
 require 'rubocop/cop/mixin/configurable_max'
