+++ lib/rubocop/cop/style/and_or.rb
@@ -5,6 +5,8 @@ module Rubocop
     module Style
       # This cop checks for uses of *and* and *or*.
       class AndOr < Cop
+        include AutocorrectUnlessChangingAST
+
         MSG = 'Use `%s` instead of `%s`.'
 
         OPS = { 'and' => '&&', 'or' => '||' }
@@ -28,20 +30,6 @@ module Rubocop
                         :operator,
                         format(MSG, OPS[op], op))
           end
-        end
-
-        def autocorrect(node)
-          c = correction(node)
-          new_source = rewrite_node(node, c)
-
-          # Make the correction only if it doesn't change the AST.
-          @corrections << c if node == SourceParser.parse(new_source).ast
-        end
-
-        def rewrite_node(node, correction)
-          processed_source = SourceParser.parse(node.loc.expression.source)
-          c = correction(processed_source.ast)
-          Corrector.new(processed_source.buffer, [c]).rewrite
         end
 
         def correction(node)
